Fix bitops.scm API
`clear-bits' was missing entirely from the public api and there were
both `or-bits' and `set-bits'. The procedure in the implementation was
actually called `or-bits' and used as such in cdce72010-prg.scm, but
that is the wrong name for the procedure, since it doesn't just "or"
bits into the old value, but clears the range of bits first, too. That
is not ORing, that's setting. Renamed accordingly.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>