Explicitly set serial device attributes
The attributes are not always auto-sensed correctly, which usually
causes lengthy (yet unjustified) debugging. This patch uses termios to
setup the serial device as 8N1 at 19200 baud.

It would be nice to have this configurable from scheme, but this is
enough for now, because this is exactly the speed the current target
device is using.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>