Rename the project to chip-remote
This should be a chip-agnostic solution rather than a single-chip one.

There still is a lot to do: Much of the C-Code still implements cdce72010
specific things, like the address-readback hardware-bug of the device; or
the eeprom-write and write-locked routines.

The C part really should only implement the bare minimum for implementing
the serial protocol outlined in my related paper.

The scheme code should be even  better divided: scheme/brand/device/tables.scm

The test-suite should probably be split into tests for the C core and
tests for the scheme extensions, like `bitops.scm', and tests
for device specific code.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>