fw: Add port implementation
This is a first attempt. It is not entirely clear whether this is
sufficient, yet. It currently works with separate read- and write-
addresses. Not sure how portable that's going to be...

The port configuration has to be malloc()ed. I don't see a way around
that if the systems is to be truly configurable.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>