tests: Add test-suite driver for firmware tests
The firmware's reactions to input should arguably be tested. We already
have a simulator-like build-setup in place, so we use that.

This avoids hand-crafting large here-docs by generating large parts of
the protocol on the fly. It sends a line to the simulator, then reads
two lines from it (the simulator always echos back the requests it gets
before producing replies). Then it compares that with what it thinks to
expect and if that differs it signals failure. Otherwise the test
passes.

The driver produces TAP output so, it fits into our current test-suite
rather nicely.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>