scheme: Add register-map abstraction
Configurable devices usually contain one or more configuration
registers, that encode the actual configuration information.

This leads to a lot of tedious, error-prone and similar code, that
nobody wants to write. So with this, you can do:

  (define-register-map xyz123
    (#x0 (default-value #x54)
         (contents (m-divider 0 4)
                   (n-divider 4 4))
    (#x1 (default-value #xd0)
         (contents (power-sabe 0 1)
                   (reference-divider 4 4))))

...which is a direct transcription of a register-map into scheme code.
The `define-register-map' will then expand into all that code, that you
didn't want to write yourself.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>