scheme: Syntactic sugar ‘chain-fncs’
Register-modifying functions in chip-remote look either like this:

  (enable-foobar register-value)

or like this:

  (set-this-thing register-value args ...)

What they have in common is, that the register-value they modify is
handed into the functions as the first argument.

To apply more than one such modification, you end up with nested
functions calls like this:

  (enable-foobar
    (set-this-thing
      (enable-bazquuz register-value)
      'argument))

The more modifiations you want to chain, the less readable the
expressions get. The introduced macro allows you to write this:

  (chain-fncs register-value
              enable-bazquuz
              (set-this-thing 'argument)
              enable-foobar)

And that expands into the same nested expression, you saw a bit up. If
the order of operation matters, ‘chain-fncs' in such a way that the
first function, you hand into the macro, becomes the innermost function
call (it is therefore evaluated first).

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>