scheme: Add `=>' macro to register-map module
To implement register-decoding, I need to be able to refer to variables
in the macro call. To do that, this macro can be used like this:

  (define-register-map foobar
    (#x0 (contents (gain 0 8 #,(=> gain-map)))))

That'll put in a variable reference in place of the "#,(=> ...)"
expression. It's ugly to have user supply the unsyntax call in there,
but I couldn't work out a way to do this more elegantly.

This might change in the future.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>