scheme: (... ads4149 program) Write unhandled functions
First things first.

These are exactly the reason why we can't generate second level
functions completely automatical. Well, we could make the generation
script smarter, but eventually, a chip will come around that does
something that breaks things.

This solves the problem a bit differently to the old manually written
code, but I like it better this way, so... there.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>