scheme: (... ads4149 program) Consistency
Sometimes, the generation script might get confused. Like with this
bit field. It actually works like a single bit with enable/disable
semantics, but since it's two bits wide, the generation yields
something else.

To be consistent with the API of other fields with enable/disable
semantics, let's rewrite this code like this.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>