scheme: Rework register-map a little
The #,(...) entries were bad to begin with. They are gone now. That
however, comes at the prize of doing some of the register map expansion
at run-time, which I'm not happy with, but a solution will present
itself, I guess.

This makes the register-map hold data they removes a few ugly hacks
from the decode module.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>