scheme: Add ‘define-register-interconns’ macro
The information in any given register of a chip may be connected to
the information of another item in another register in the chip.

This makes level-2 and level-3 code generation a bit harder, the
accurate decoding of single registers impossible in the general case
and the decoding of an entire chip's register map harder.

The introduction of this script should help with two of the three
issues. You still just can't decode a single register accurately,
if the way the decoding has to happen depends on the value of a
setting in some other register...

But with the other issues, this should help.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>