Add a new register table compiler: front end
The idea is to have a "register.table" file that consists of forms
valid in the realm of a domain-specific language, designed to
describe register tables of configurable chips. Using that source
languange then allows to write compiler back-ends to generate first,
second and third-level access code; encoders, decoders or back-ends
that do entirely different things like compiling to micro-controller
C code.

This implements most of the machinery for the front-end as well as
support for the "register" form, which is what is required to
write a compiler back-end that generate the first level of chip
access.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>