Add new device CDCE72010
This is going to be a reimplementation of the CDCE72010 support. The old
interface is in (chip-remote legacy ti cdce72010). This new module uses
the compiler framework and will not be API compatible with the old
module.

It is very much a sand box for figuring out how to use and develop the
DSL and its compiler framework.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>