Latest work in progress from previous implemenation plan
This plan was a replaced by a revised library layout that better leverages the
host language, rather than implementing a new one. This commit is done for com-
pleteness' sake. The compiler code will be phased out shortly.

All existing device support libraries will be ported to the new API. The lagacy
namespace will be nuked as well.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>