Add a minimal interpreted language
This may seem heavy-handed, but it has its purpose: Since chip-remote's goal is
not only to play with device configurations in Scheme, but also to be able to
compile down to other target languages such as C or C++, it is important to
keep information about source code, in case custom solutions such as encoders,
decoders or validators need to be implemented.

This could be done by quoting Scheme expressions and evaling those, but Scheme
is quite powerful and its full specs are hard to compile to C or something
similar. This interpreter has quite a limited feature set and its evaluator
yields Scheme values, so its output can be used within the host language.

A new type <evaluation> is introduced that keeps the source-code for the
interpreter alongside the value yielded by the evaluator. We'll store this, for
instance, in custom #:semantics blocks so compiling to other languages is
possible.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>