Start new decoder framework (very basic for now)
The idea is to hand one of our device-describing data types and a matching
datum in, find the right decoder and go to town.

This will not contain a pretty printer. That is going to be a post processor
built on top of this. I wonder how much meta-information is required for that
post-processing to be powerful enough. If the structure of the data and the
data description matches (and it has to!) the post processor could zip the
description and the data and decode like that. It would have all the available
information to work with, which seems elegant.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>