Fix register/item decoder and their tests
First issue here was, that the new REPL decoder annotates its result with a
name. That one is easy to fix by amending the expected values in the tests.

The second problem was, that the register decoder didn't hand the correct value
for the item to the item decoder. Instead it passed the register value down.
And that obviously breaks for most cases.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>