Add a basic unit system
The goal is to be able to attach unit information to values and convert between
them with ease. The system should also be extensible: Make up new dimensions
and units and work with them naturally: Convert, combine, make calculations.

This first implementation implements a couple of fundamental dimensions and
units as well as SI prefixes (such as kilo and milli). Introducing new units
based on fundamental units is supported. But there is lots to do still. Adding
new dimensions (like electrical-voltage for example) is not yet supported.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>