reg: Allow embedding evaluations into #:content ... data
Sometimes, it will be useful to drop in evaluated items into a register, to
insert items that are similar across registers with lots of similar meta data,
like address items or reserved items.

This was already doable, but you'd have to interupt #:contents by #:contents*
and so on. Now you can do this:

(generate-resister
  #:contents
  (=> (address 0))
  (foo 2 3)
  (=> (reserved 5 2))
  (pause? 7 1)
  (=> (reserved 8 4))
  (divider 12 4))

…which is a step into the right direction.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>