Add register-window data type
The idea here is to be able to look at a part of a register. For example if you
have a large registers with tons of items, but you're really only interested in
one specific chunk out of this register, then this data type can help. Take a
register, specify an offset and a width in bits and you got your window into
that register.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>