Add an example that performs default setup in a ADF4169 PLL
This script connects to a remote firmware, configures a port to bit-bang SPI,
configures that SPI to match the ADF4169's specs and then transfers the chip's
default configuration (as per the adf4169 module) into the chip.

This script will serve as a blue-print to identify further abstraction to
simplify the application of the library.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>