Pretty print a couple of types by default
With chip-remote, data is made up of semantics in items in register in
register-maps in page-maps in devices… This high level of nesting makes
printing such values rather tough to comprehend, because the default
repl-printer causes all data to be printing in one line.

The plan is to add pretty-printing similar to (ice-9 pretty-print), while
leveraging the additional knowledge about the specific types.

This does not add pretty printing for all chip-remote types, yet.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>