fw: Add new embedded protocol parser
This is the beginning of a new remote-firmware, that will be based upon the
Zephyr operating system. This adds all infrastructure from an incoming protocol
buffer to a callback structure that includes a spec of required/allowed
arguments. This allows early error checking so the actual command handler has
to perform as little boiler plate actions as possible.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>