fw: Configure kernel properly with native posix build
This enables two UARTs that are/will be used in the native build. This allows
the library to connect to the firmware as a unix process and talk to it as if
it where running on actual hardware.

To play with this, you can also use programs such as picocom:

  picocom --echo --omap crlf --imap lfcrlf /dev/pts/14

with "/dev/pts/14" being the device file the unix program prints to stdout upon
boot for UART_0.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>