fw: Add configuration for Nucleo-F767ZI board
This is the board the old firmware targeted; this doesn't work yet, because
here main.c just spawns the protocol processor thread, but nothing that would
feed it. That'll be added shortly.

For now, to build this on debian linux with the GNU ARM embedded toolchain in-
stalled, use this:

  mkdir build
  cd build
  export GNUARMEMB_TOOLCHAIN_PATH=/usr
  export ZEPHYR_BASE="$PWD/../../kernel/zephyr"
  export ZEPHYR_TOOLCHAIN_VARIANT=gnuarmemb
  MODULES="$PWD/../../kernel/modules/cmsis"
  MODULES="$MODULES;$PWD/../../kernel/modules/hal_stm32"
  cmake -GNinja                     \
        -DBOARD=nucleo_f767zi       \
        -DZEPHYR_MODULES="$MODULES" \
        ..

Should probably figure out how to do this using west.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>