fw: Preliminary adaption for the STM32 Nucleo-F767ZI board
This is a pretty powerful evaluation board; I'm just blinking a LED for now and
connect the board's USB port as a CDC ACM port to the chip-remote machinery.

The board responds properly now. What's missing is a port definition and a
hardware transmit callback so we can see actual pins wiggling. But this is the
foundation.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>