Remove state from protocol
Having the explicit Idle/Active/Multiline state was a mistake when defining the
ASCII protocol used by chip-remote. It doesn't add functionality and heavily
increases the complexity of implementaion. The ideas behind these states were
to have a way to check whether or not a serial connection works (baudrate, pa-
rity and stop bits, things like that). And having a request that causes a known
reply is a good thing to have. But moving this initialisation conversation into
its own state is not. We'll keep both commands (HI and BYE), without having a
separate state.

The second idea (multiline state) was to keep the allowed length of protocol
buffers low. This is still a good idea, but the only buffer that's important to
keep small is the receive buffer in the remote firmware. The receive buffer in
the controlling client is likely rather large. After all, the whole idea behind
chip-remote is to connect configurable chips to a desktop computer that allows
for more comfortable control and processing. Therefore, we're now allowing
arbitrarily large replies from the remote firmware. Requests still need to
remain short in order to keep the firmware's receive buffer small. But allowing
this remove the need for the multiline state.

And with that, both additional states (multiline and idle) can be removed. The
protocol therefore is still line-based, while each reply completely resolves a
request.

This change still needs to be reflected in a specification update.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>