Reimplement approach to combinations
A combination is a meta-item, that's made up of multiple items to form some-
thing bigger. This approach is (hopefully) more general and simpler then the
previous attempts.

There are two main functions to its API:

  - combination-assemble
  - combination-partition

The former takes a device and its value and forms a combination value that can
be decodes via user-supplied semantics. The latter takes a combination's value,
a device and returns a script, that — when executed — applies changes to all
items required for the device to reflect the desired combination's value.

This needs to be applied to the device abstraction and the modification and de-
code modules, so that from a user's point of view there will be little diffe-
rence between primitive items and combination items.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>