Remove the need for sem-ish? from combination runner
Before revamping the whole semantics idea, the allowed value for a semantics
slot was either a semantics value, or a procedure accepting a width argument
that, when called, produces a semantics value.

Now that all semantics codecs require a width argument, this is not required
anymore and this function is rendered useless.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>