Remove old remote-fw implementation
The new firmware code is in ‘firmware’. That code is based upon the Zephyr real
time ecosystem, which brings its own board abstraction, which is nice, because
we don't have to invent our own anymore. It also deals nicely with integrating
vendor libraries, which means no more included third-party code.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>