Add annotation to each state log value
It is useful to know what created a new state for a device. May be a read, a
write or a local set-application, at least. This will give us a pretty complete
transaction log with a device, which should be useful while debugging.

A device-value diff and a decoding diff would be useful right about now.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>