Add function to minimise required transfers for modification script
When manipulating data in memory, the number of changes made are not too
important, as long as the result is correct. When transferring data into a
device, that changes radically. Here you only want to perform as few transfers
as possible to achieve the equivalent result.

This adds a function that can reduce the number of required transmissions from
a script as constructed by chain-modify-script to a minimum.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>