Add a few improvements to the scratch pad file
More useful modules for experimentation and the ‘ignore’ function, which allows
a user to only observe the side-effects of a procedure call. This is useful to
remove the return values from REPL output, which can be tough comprehend be-
cause chip-remote builds pretty large data structures. Often too large for REPL
use.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>