tests: Add utilities to interact with natively built firmwares
This is groundwork to talk to the Zephyr application build against one of the
native targets. When starting, the application notes the file names of the
device files it connected its UART ports to. We need UART_0 to connect our chip
remote client library to.

After that we can talk to it.

While already reading the application's stdout, we can keep reading it and make
the firmware emit internal data as its running. That way it is possible to test
more aspects about the system.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>