tests: Add a simple system test that runs native firmware build
There are some parts to this that make writing these kinds of tests concise:

  (make-test-io)     ;; Creates a handle for all aspects of the
                     ;; running firmware.

  (require (native-firmware-built?))
                     ;; Causes guile-tap to SKIP the whole test
                     ;; bundle in case the native firmware build
                     ;; could not be found.

  (boot-fw! tio)     ;; Boots the firmware and updates ‘tio’
                     ;; accordingly.

  ($ tio)            ;; Shortcut to fetch chip-remote connection
                     ;; from ‘tio’.

  (kill-fw! tio)     ;; Terminates the firmware process.

This test starts the firmware, connects to its chip-remote terminal on UART_0
and tries to perform the ‘hi’ and ‘bye’ requests. The procedures running those
throw exceptions when something goes wrong with that, so pass-if-no-exception
is a proper test.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>