fw: Add s-expression parser
The plan is to add instrumentation to the native posix build of the firmware.
That way we can tell the firmware what kind of behaviour we're expecting from a
device attached to it, hence being able to test huge parts of the system in
orchestration.

We'll be using s-expressions for communicating instrumentation, because it is a
fairly simple, yet powerful system. Since we'll be using this *only* for the
native build, we're not tied to the same constraints as is the rest of the
firmware. As a result this parser uses dynamic memory quite liberally.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>