tests: Implement ways to suspend system test to make debugging easier
System tests using zephyr's native builds are nice, because you can run tests
with the real firmware code and the real OS as if it was a program on your
desktop, no hardware required. And with our little instrumentation protocol we
can control the paths of execution that are taken pretty nicely.

But what if you encouter a failing test? It would be nice if you could attach a
gdb to the given PID and dig into program state. Right? Right. But if you
tried, you'd quickly see that stuff breaks because you'll quickly run into
timeouts everywhere. In normal execution, these timeouts are good, because a
program should respond in finite time. With debugging virtually all of these
timeouts may trigger, just because you're in a breakpoint.

This extends ‘boot-fw!’ to also parse the command line parameters passed to the
test program. If the first parameter is "--debug", the procedure will tune all
timeouts it knows about to infinity and halt the excution of the test program
as soon as it has determined the firmware's PID:

    # Booting native firmware: …/chip-remote/native-fw/zephyr/zephyr.exe
    # Chip-Remote Terminal at: /dev/pts/22
    # Instrumentation Terminal at: /dev/pts/21
    # S-Expression Interface up.
    # Registering firmware PID: 17545
    # Debug Mode (PID: 17545): Press ENTER to continue!

This allows you to attach a gdb to the running firmware process:

    % gdb -p 17545 ./native-fw/zephyr/zephyr.exe

When you're done setting up your debugger you can press enter with the
suspended test program and the tests will run with your debugger observing.

Finally, if you've got a long test program and the issue you're after is
happening at a test late in the script and you're not interested in wading
through all execution that doesn't concern you, here's a way of manually
inserted a suspension point:

    ;; This will cause boot-fw! not to wait for user input
    ;; after determining the firmware's PID.
    (boot-fw! tio #:suspend-execution? #f)
    …lots of test-code…
    ;; Finally when you'd like to cause the test program to suspend:
    (debug-fw! tio)

You may use ‘debug-fw!’ as often as you like.

Using "--debug" also causes ‘kill-fw!’ to not actually kill the firmware
program's PID, since maybe you'd like to keep poking at your program's state
even after all tests have executed.

Therefore, using "--debug" makes it your responsibility to kill the firmware
process. Using gdb, you can just Ctrl-C and issue gdb's ‘kill’ command. That'll
do the trick.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>