scheme: API cleanup and value reorganisation
The addressing API got out of hand on a bunch of places with special purpose
code all over the place. The addressing APIs are now uniform across all data
types:

- <type>-address: Allows the convenient reference of an item within a larger
  structure.

- <type>-address:register: Like the previous function, but returns the register
  that the addressed item resides in.

- <type>-canonical: Return the canonical address of an item within a bigger
  structure.

- <type>-ref: References an arbitrary substructure within a bigger structure.

By design, that allows the implementation of the <type>-address function as

  (apply <type>-ref (apply <type>-canonical ...))

That goes similar with the :register variant by just dropping the item index
from the canonical address.

While doing this redesign, the previous values for more complex structures was
found to be lacking. Previously, the value for a system with a page-map con-
taining two registers each might have had a possible value such as:

  ((0 0)(0 0))

That is rather compact and neat, but it also burdens the implementation to do
lots of backward calculation to find out which integer belongs to what regi-
ster, given that addresses don't have to be contiguous or even ordered in spe-
cifications. Now values for larger structures resemble that structure much more
than they did before. Now a value for the previously described structure could
be:

  ((0 (0 . 0) (10 . 0))(100 (12 . 0) (34 . 0)))

Here the page addresses would be ‘0’ and ‘100’ with registers being addressed
by ‘0’ and ‘10’; and ‘12’ and ‘34’ respectively. This is much less compact but
addressing the values in question can be done using the same canonical addres-
ses used to address items in bigger structures.

There will probably be some cleanup required still. And the modify module could
certainly use a little more of the new APIs.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>