fw: Add generic parser for the new procotol spec
The old parser is a little inflexible and makes things awkward with the plans
to forward more functionality to implementation specific backends. This parser
generates a list of tokens that are one of symbols, booleans, numbers and key-
value pairs.

The new parser still works in completely static memory without any dynamic al-
locations.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>