proto,slip: Discard data between frame-end and frame-start delimiters
Finding the start-of-frame octet should discard all previously gathered data,
since all data between end-of-frame and start-of-frame is to be considered
garbage. It is possible that by sheer luck, the garbage contains a start-of-
frame octet as well. And everything up to the real start-of-frame is therefore
garbage as well. It is also possible that the garbage contains an end-of-frame
delimiter too. In that case, there is nothing we can do.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>