io: Make reading from firmware more robust
It appears that read-line may return early in some cases. Not sure why, but
maybe it is some kind of optimisation. In any case, with slow ports such as
serial ports, sometimes I see the function return data that was not yet
delimited by a newline. The rest of the line that was prematurely returned is
then returned in the next read. That breaks logic in a request/response based
protocol severely.

This implements the desired logic manually by always reading all available
data, inspecting for newline content and reading more if such a character could
not be found. If a read reveals data after a newline character, that data is
pushed back. This shouldn't happen in this request/response scheme, but it's
good form to do it anyway.

I've tried this for a while and the multi-read-read operation happens every now
and then and the new code handles the situation correctly.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>