fw: Port to nucleo l476rg board
This is a compact development board of st micro featuring an STM32L476, an
ultra-low-power Cortex-M4.

The board does not feature a network port or an additional USB link. But it
comes with an integrated stlink-v2 on-board, that has the ability to offer one
serial port (usart2 in this board's case) through the debugger link.

Normally, zephyr puts its console on there, but in order to have convenient
chip-remote functionality, we're rerouting console to usart1 (PA9, PA10 on the
boards IO headers). If you need zephyr's console, connect a UART transceiver to
those pins.

When connecting this board you'll roughly see this:

[112092.256072] usb 1-4: USB disconnect, device number 6
[112094.455793] usb 1-4: new full-speed USB device number 7 using xhci_hcd
[112094.609655] usb 1-4: New USB device found, idVendor=0483, idProduct=374b, bcdDevice= 1.00
[112094.609660] usb 1-4: New USB device strings: Mfr=1, Product=2, SerialNumber=3
[112094.609662] usb 1-4: Product: STM32 STLink
[112094.609663] usb 1-4: Manufacturer: STMicroelectronics
[112094.609665] usb 1-4: SerialNumber: 066BFF333535554157153641
[112094.668230] cdc_acm 1-4:1.2: ttyACM0: USB ACM device

With that you can connect to the chip-remote firmware via /dev/ttyACM0 using
115200bd 8N1. To play with the firmware via picocom, you can do this:

    % picocom -q --echo --omap crlf --imap lfcrlf -b 115200 /dev/ttyACM0
    hi
    Hi there, stranger.
    transmit h#123
    h#0
    bye
    Have a nice day.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>