fw: Upgrade firmware operating system to zephyr version 3.2.0
Lots of things have changed in zephyr: The USB subsystem, the include path
setup requiring ‘zephyr/’ as a prefix, lots of device-tree related aspects.
Also, we can't presently use the variant of newlib shipped by debian, since it
does not enable the relocatable-locking feature.

So this also adds a dependency upon the picolib zephyr module. This libc has a
similar feature set as newlib, but zephyr builds it as part of its compilation
system, hence enabling the features it requires.

This port wasn't properly tested yet, but at least building the firmware is
fixed again.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>