Actually fix decoding with new page/reg-map values
The new data structures actually map addresses to values, instead of implicitly
doing this mapping. The upside of this is, of course, that you can tell from
seeing a value, which value belong to what address, without needing to matching
the value to its semantic description.

With decoding framework was not updated to do deal with this. We need to hand
values that are not decorated by addresses to the lower layers of the decoder.
This therefore refactors the map generation into a function that has an
optional transform parameter. If given a pagemap value such as

    (0 (0 . 16))

…the decoder needs to pass (0 . 16) to the register-map decoder layer, and the
register map decoder needs to pass 16 to the register decoder layer. So as it
turns out, the transform parameter happens to be ‘cdr’ in all cases. We're
leaving it in for clarity, and the ability to adjust for future changes.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>