Implement loading values into entities larger than an item
There are use-cases where it might be helpful to load complete register values
into a device-value (or page-map or register-map), or even one of the mapping
values.

While replace-register-value existed, it didn't integrate nicely into the set!
API of the commander interface.

This is implemented by prefixing the ADDRESS parameter of a change spec by the
‘#:load’ keyword, so you can do: (commander 'set! '(#:load #f 15 #xffffffff)),
to load #xffffffff into the register at page-address #f and register address 15
directly.

This enables all other features of the commander interface too, including value
histories.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>