cmd: Add address query helpers
When looking at larger devices, it is useful to decode parts of the device
instead of everything. When doing this, it is sometimes useful to look at a
larger unit, that contains said item. To be able to do this, it is useful to be
able to get the canonical address of the item to see the addresses of the
enclosing units (registers, register-maps).

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>