fw: Strip old protocol implementation
This is in preparation of using ufw's register implementation as the next
generation protocol for the firmware.

There are still some remnants of the old implemenatation in the peripheral in-
terface implementations in "ifc", which is currently just not compiled. There
might be one or the other piece of infrastructure that may be possible to re-
use. But most of that will likely also go away eventually.

The system tests that depend on the chip-remote protocol to be available are
disabled for the time being.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>