fw: Unify firmware initialisation from init-*.c into main.c
Zephyr's subsystems (devicetree and kconfig) allow lots of compile time
introspection. This allows for dumbing down the cmake side (which is always a
good thing), and reduce code duplication for initialisation.

The new file "main.c" is a little bit littered with #ifdef statements, but it
is still relatively straight forward. So we're not putting in another layer of
abstraction in.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>