Add records data type from Guix
Chip-remote uses plists to create a richer way to work with data types. But the
implementation is not nice at all, and it is really difficult at the usage
site.

The record type from GNU Guix looks really well thought out.

Replacing this is a bit of an undertaking, but I think it is worth it.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>