Port all device descriptions to records* API
Most of these were ported manually. But the two massive dw1000/dw3000 register
modules were refactored using a script. It is checked into the repository as
"tools/dwx000-register-refactor.scm", in case it should be needed again.

This passes the whole test-suite again, but there might still be some remnants
around to screw with some modules.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>