Rework standalone operation
I like GNU Guix, and I'd like to use it to make development environments with
it. That's where manifest.scm comes in. It can be used with guix shell to make
an environment that is suitable for working on chip-remote.

Guix brings guile-tap and guile-termios, cool! But test-dispatch and mmh are
not part of it. None of these are available on other systems like Debian. So
you'd have to install them into the system before. That's where the old stand-
alone system came in.

The new is written in Scheme and can test of the host has any of the dependen-
cies available. If so, it skips adding them to standalone. This works better
with things like Guix or hosts that have these things installed. The new code
also allows adjusting the URL from which dependencies are cloned.

The Guix manifest currently works well for the library. For the firmware, there
is room for improvement. But it's a start.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>