Redesign device access (again)
Now a device implementation does not directly talk to a chip via the
chip-remote firmware anymore. Instead, it implements an API to produce data
that, when transmitted into a chip, achieves read and write operations.

Additionally, there is now a device-bus type, an abstraction of how to transfer
data into a chip. Such busses are SPI and I2C.

In combination, device-access and device-bus perform transactions with real
devices — or fake devices, because you can just as well implement a bus purely
in software. The (chip-remote access) module implements a simple fake-spi bus.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>